Revert r25824:25875 from the core.svn.wordpress.org repository.
These commits were accidentally re-synced commits from develop.svn.wordpress.org due to a race condition. Thankfully, the history of this repository matters fairly little. It also happened only for trunk.


git-svn-id: http://core.svn.wordpress.org/trunk@25876 1a063a9b-81f0-0310-95a4-ce76da25c4cd

+++ wp-admin/includes/update-core.php
@@ -561,6 +561,7 @@ $_old_files = array(
 // 3.7
 'wp-admin/js/cat.js',
 'wp-admin/js/cat.min.js',
+'wp-includes/js/tinymce/plugins/wpeditimage/js/editimage.min.js',
 );
 
 /**
@@ -894,7 +895,8 @@ function update_core($from, $to) {
 	do_action( '_core_updated_successfully', $wp_version );
 
 	// Clear the option that blocks auto updates after failures, now that we've been successful.
-	delete_site_option( 'auto_core_update_failed' );
+	if ( function_exists( 'delete_site_option' ) )
+		delete_site_option( 'auto_core_update_failed' );
 
 	return $wp_version;
 }
